py/browseruse/tutorial/pyproject.toml
[project]
name = "tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "browser-use>=0.1.23",
    "python-dotenv>=1.0.1",
]
